feat: global GMM for stable archetypes, LEFTIES/RIGHTIES labels, zoom fix
Switch from per-year independent GMMs to a single global GMM across all
years (2015-2025). Archetypes are now consistent year-over-year — same
pitcher with a stable pitch mix stays in the same cluster. Real evolution
(e.g. dropping a pitch, adding a cutter) shows as actual archetype changes.

- 26 clusters (15 RHP + 11 LHP) with soft GMM probabilities per season
- Rewrote export_frontend.py to use global model outputs
- Re-exported all frontend JSON with consistent R_/L_ cluster keys
- Regenerated hitter matchup data (06 + 07) with new cluster IDs
- Added LEFTIES / RIGHTIES banners floating above each side of town
- Zoom sensitivity 2.5x (0.8 → 2.0), closer min (1.5), further max (80)
- Added Wormburner + Heater-Heavy archetype descriptions
- Prevented 06_hitter_vs_cluster from overwriting pitcher_seasons.json

+++ pipeline/05_cluster_naming.py
@@ -272,7 +272,7 @@ def main():
         if len(members) == 0:
             continue
         feature_cols = [f for f in features_used if f != "is_rhp" and f in members.columns]
-        coords = members[feature_cols].fillna(0).values
+        coords = members[feature_cols].fillna(0).values.astype(np.float64)
         if len(coords) < 2:
             medoid_rows[cid] = members.iloc[0]
         else:
